requirements.txt
fastapi
uvicorn
websocket-client
gspread
loguru
requests
python-dotenv
